Fast and Air
Changed Wind speed to be interleaved in the altitude array to reduce the number of copy opperations, speed increase by about 20x!

Added an icon and title to the gui

Atmospheric temperature and pressure are now initalized on the first timestep if it isn't already. (init stored in world data so it gets loaded in together if saved)