Fixed negative kelvin...
Due to a bug where air mass was incorrectly calcualted for high temperatures mass could become negative.
calculation was based on air density which is now more accurate and factors in humidity and pressure.

Next up, cooling so the world doesn't burn.

It's been burning, since the world's been turning.